+++ spec/rspec/expectations/configuration_spec.rb
@@ -157,6 +157,24 @@ module RSpec
           configure_syntax(@orig_syntax)
         end
 
+        it "warns when should syntax is choosen" do
+          expect(RSpec).
+            to receive(:deprecate).
+            with("Expectations syntax configuration",
+                 :call_site=>nil,
+                 :replacement=>"the default `expect` syntax").
+            at_least(:once)
+
+          expect(RSpec).
+            to receive(:deprecate).
+            with("`:should` Expectations syntax",
+                 :call_site=>nil,
+                 :replacement=>"the default `expect` syntax").
+            at_least(:once)
+
+          configure_syntax :should
+        end
+
         it 'can limit the syntax to :should' do
           configure_syntax :should
           configured_syntax.should eq([:should])
@@ -189,37 +207,6 @@ module RSpec
 
           configure_syntax :expect
           configure_syntax :expect
-        end
-
-        describe "`:should` being enabled by default deprecation" do
-          before { configure_default_syntax }
-
-          it "warns when the should syntax is called by default" do
-            expected_arguments = [
-              /Using.*without explicitly enabling/,
-              { :replacement => "the new `:expect` syntax or explicitly enable `:should` with `config.expect_with(:rspec) { |c| c.syntax = :should }`" }
-            ]
-
-            expect(RSpec).to receive(:deprecate).with(*expected_arguments)
-            3.should eq(3)
-          end
-
-          it "includes the call site in the deprecation warning by default" do
-            expect_deprecation_with_call_site(__FILE__, __LINE__ + 1)
-            3.should eq(3)
-          end
-
-          it "does not warn when only the should syntax is explicitly configured" do
-            configure_syntax(:should)
-            RSpec.should_not receive(:deprecate)
-            3.should eq(3)
-          end
-
-          it "does not warn when both the should and expect syntaxes are explicitly configured" do
-            configure_syntax([:should, :expect])
-            expect(RSpec).not_to receive(:deprecate)
-            3.should eq(3)
-          end
         end
 
         it 'can re-enable the :should syntax' do